Introduce test condition for element connector.
Deprecate functions in TestTool.
