+++ californium-core/src/test/java/org/eclipse/californium/core/network/GroupedMessageIdTrackerTest.java
@@ -18,16 +18,16 @@ package org.eclipse.californium.core.network;
 
 import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.*;
-import static org.eclipse.californium.TestTools.inRange;
 import static org.eclipse.californium.core.network.MessageIdTracker.TOTAL_NO_OF_MIDS;
+import static org.eclipse.californium.elements.util.TestConditionTools.inRange;
 
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.eclipse.californium.CheckCondition;
-import org.eclipse.californium.TestTools;
 import org.eclipse.californium.category.Small;
 import org.eclipse.californium.core.network.config.NetworkConfig;
+import org.eclipse.californium.elements.util.TestCondition;
+import org.eclipse.californium.elements.util.TestConditionTools;
 import org.eclipse.californium.rule.CoapNetworkRule;
 import org.eclipse.californium.rule.CoapThreadsRule;
 import org.junit.ClassRule;
@@ -120,7 +120,7 @@ public class GroupedMessageIdTrackerTest {
 		}
 
 		final AtomicInteger mid = new AtomicInteger(-1);
-		TestTools.waitForCondition(timeLeft, 100, TimeUnit.MILLISECONDS, new CheckCondition() {
+		TestConditionTools.waitForCondition(timeLeft, 100, TimeUnit.MILLISECONDS, new TestCondition() {
 
 			@Override
 			public boolean isFulFilled() throws IllegalStateException {
