+++ californium-core/src/test/java/org/eclipse/californium/core/network/NullMessageIdTrackerTest.java
@@ -18,7 +18,7 @@ package org.eclipse.californium.core.network;
 
 import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.*;
-import static org.eclipse.californium.TestTools.inRange;
+import static org.eclipse.californium.elements.util.TestConditionTools.inRange;
 import static org.eclipse.californium.core.network.MessageIdTracker.TOTAL_NO_OF_MIDS;
 
 import org.eclipse.californium.category.Small;
