+++ californium-core/src/test/java/org/eclipse/californium/core/test/lockstep/DeduplicationTest.java
@@ -47,6 +47,7 @@ import org.eclipse.californium.core.network.stack.ReliabilityLayerParameters.Bui
 import org.eclipse.californium.core.test.CountingMessageObserver;
 import org.eclipse.californium.core.test.MessageExchangeStoreTool.UDPTestConnector;
 import org.eclipse.californium.elements.rule.TestNameLoggerRule;
+import org.eclipse.californium.elements.util.TestConditionTools;
 import org.eclipse.californium.rule.CoapNetworkRule;
 import org.eclipse.californium.rule.CoapThreadsRule;
 import org.hamcrest.Matcher;
@@ -232,10 +233,10 @@ public class DeduplicationTest {
 
 	private void assertHealthCounter(final String name, final Matcher<? super Long> matcher, long timeout)
 			throws InterruptedException {
-		TestTools.assertCounter(health, name, matcher, timeout);
+		TestConditionTools.assertStatisticCounter(health, name, matcher, timeout, TimeUnit.MILLISECONDS);
 	}
 
 	private void assertHealthCounter(String name, Matcher<? super Long> matcher) {
-		TestTools.assertCounter(health, name, matcher);
+		TestConditionTools.assertStatisticCounter(health, name, matcher);
 	}
 }
