+++ californium-core/src/test/java/org/eclipse/californium/core/test/lockstep/ObserveServerSideTest.java
@@ -52,7 +52,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.eclipse.californium.CheckCondition;
 import org.eclipse.californium.TestTools;
 import org.eclipse.californium.category.Medium;
 import org.eclipse.californium.core.CoapResource;
@@ -66,6 +65,8 @@ import org.eclipse.californium.core.test.ErrorInjector;
 import org.eclipse.californium.core.test.MessageExchangeStoreTool.CoapTestEndpoint;
 import org.eclipse.californium.elements.rule.TestNameLoggerRule;
 import org.eclipse.californium.elements.rule.TestTimeRule;
+import org.eclipse.californium.elements.util.TestCondition;
+import org.eclipse.californium.elements.util.TestConditionTools;
 import org.eclipse.californium.rule.CoapNetworkRule;
 import org.eclipse.californium.rule.CoapThreadsRule;
 import org.junit.After;
@@ -623,7 +624,7 @@ public class ObserveServerSideTest {
 
 	private int waitForObservers(long timeoutMillis, final int count) throws InterruptedException {
 
-		TestTools.waitForCondition(timeoutMillis, 50, TimeUnit.MILLISECONDS, new CheckCondition() {
+		TestConditionTools.waitForCondition(timeoutMillis, 50, TimeUnit.MILLISECONDS, new TestCondition() {
 
 			@Override
 			public boolean isFulFilled() throws IllegalStateException {
