+++ scandium-core/src/test/resources/logback-test.xml
@@ -14,7 +14,7 @@
 
 	<!-- Strictly speaking, the level attribute is not necessary since -->
 	<!-- the level of the root level is set to DEBUG by default. -->
-	<root level="DEBUG">
+	<root level="WARN">
 		<appender-ref ref="STDOUT" />
 	</root>
